[BH-4] Handle loading textures, general cleanup

+++ Bullet_Hell/src/ResourceCache/ResourceFile.h
@@ -27,7 +27,7 @@ public:
 	/// </summary>
 	/// <param name="resource">The resource we want the size of.</param>
 	/// <returns>The size, in bytes.</returns>
-	virtual int get_raw_resource_size(const Resource& resource) = 0;
+	virtual size_t get_raw_resource_size(const Resource& resource) = 0;
 
 	/// <summary>
 	/// Read the resource from the file.
@@ -35,13 +35,13 @@ public:
 	/// <param name="resource">The resource to read.</param>
 	/// <param name="buffer">The buffer to load into.</param>
 	/// <returns>The size of the loaded resource, in bytes.</returns>
-	virtual int load_resource(const Resource& resource, char* buffer) = 0;
+	virtual size_t load_resource(const Resource& resource, char* buffer) = 0;
 
 	/// <summary>
 	/// Calculates the number of resources that are in the file.
 	/// </summary>
 	/// <returns>The number of resources in the file.</returns>
-	virtual const int get_resource_count() = 0;
+	virtual const size_t get_resource_count() = 0;
 
 	/// <summary>
 	/// Fetches the name of the n'th resource in the file.
