+++ Bullet_Hell/src/ResourceCache/ResourceZipFile.h
@@ -50,7 +50,7 @@ public:
 	/// </summary>
 	/// <param name="resource">The resource we want the size of.</param>
 	/// <returns>The size, in bytes.</returns>
-	virtual int get_raw_resource_size(const Resource& resource);
+	virtual size_t get_raw_resource_size(const Resource& resource);
 
 	/// <summary>
 	/// Read the resource from the file. If the resource is not
@@ -59,13 +59,13 @@ public:
 	/// <param name="resource">The resource to read.</param>
 	/// <param name="buffer">The pre-allocated buffer to load into.</param>
 	/// <returns>The size of the loaded resource, in bytes.</returns>
-	virtual int load_resource(const Resource& resource, char* buffer);
+	virtual size_t load_resource(const Resource& resource, char* buffer);
 
 	/// <summary>
 	/// Calculates the number of resources that are in the file.
 	/// </summary>
 	/// <returns>The number of resources in the file.</returns>
-	virtual const int get_resource_count();
+	virtual const size_t get_resource_count();
 
 	/// <summary>
 	/// Fetches the name of the n'th resource in the file.
